refactored server to be modular, so it can be used in server.js for dev and api.js for Netlify
