+++ outputs/web-app/v4/api-server/src/handlers/health.js
@@ -1,0 +1,12 @@
+/**
+ * Health check endpoint handler
+ */
+const healthHandler = (req, res) => {
+  res.json({
+    status: "healthy",
+    timestamp: new Date().toISOString(),
+    service: "todo-api-server",
+  });
+};
+
+module.exports = { healthHandler };
